# Getting Started with Mopuntain bike
## Available Scripts
The sport or activity of riding a mountain bike (a type of bicycle that has a strong frame, thick tires, and straight handlebars and that is used for riding over rough ground). Mountain Bicycles are design for riding rough off-road trails. They have flat or upright handlebars, and a very low gear range for pedaling up steep trails. Most mountain bikes have some type of shock absorbers or suspension.

# `The features you will get`

* Topeak Joeblow Booster floor pump
* It is very helpful to rod jump and switch another quickly
* Genuine Innovations Tackle Kit
* Mudhugger Shorty X front mudguard
* Buff Merino neckwear
* Camelbak Podium Chill water bottle